Update AddJS.js
Implemented a side bar for easier access to devices. Click on a category to both show all items in the category on the page as well as expand the list so that all the subcategories that have their own subcategories are shown. Clear the sidebar by refreshing the page. Still need to work on: showing multiple pages of results, if you click on a category a second time it expands again rather than collapsing, and the formatting and styles are still funky.